Module Sedlexing.Latin1

val from_gen : char Gen.t -> lexbuf

Create a lexbuf from a Latin1 encoded stream (ie a stream of Unicode code points in the range 0..255)

val from_channel : in_channel -> lexbuf

Create a lexbuf from a Latin1 encoded input channel. The client is responsible for closing the channel.

val from_string : string -> lexbuf

Create a lexbuf from a Latin1 encoded string.

val lexeme : lexbuf -> string

As Sedlexing.lexeme with a result encoded in Latin1. This function throws an exception InvalidCodepoint if it is not possible to encode the result in Latin1.

val sub_lexeme : lexbuf -> int -> int -> string

As Sedlexing.sub_lexeme with a result encoded in Latin1. This function throws an exception InvalidCodepoint if it is not possible to encode the result in Latin1.

val lexeme_char : lexbuf -> int -> char

As Sedlexing.lexeme_char with a result encoded in Latin1. This function throws an exception InvalidCodepoint if it is not possible to encode the result in Latin1.